Properly handle program already running Misc minor changes

+++ manual/parallel-launcher.install/tools/chocolateyuninstall.ps1
@@ -9,20 +9,29 @@ $packageArgs = @{
   silentArgs   = '/VERYSILENT /SUPPRESSMSGBOXES /NORESTART /SP-'
 }
 
-# It seems chocolatey waits to run the auto uninstaller until after this script,
-# which is unfortunate.
-# Start-Sleep -Seconds 10
+# Prevent uninstall if retroarch (which Parallel Launcher uses) is running, to ensure
+# no progress is lost
+# This cannot be moved to chocolateybeforemodify.ps1 unless this feature is added:
+# https://github.com/chocolatey/choco/issues/1731
+Start-CheckandThrow "retroarch"
 
-$uninstallerRunning = Get-Process "unins000*"
+# It seems chocolatey waits to run the auto uninstaller until after this script has finished,
+# rending this block actually useless. Maybe there is a workaround that I have not found.
+# It's not a huge deal, but chocolatey will report successful uninstallation before the uninstaller has fininshed.
+# Start-Sleep -Seconds 30 # Uncomment to see that the auto-uninstaller runs after this script.
+$uninstallerRunning = Get-Process -Name "unins000*"
 if ($uninstallerRunning) {
   Write-Output 'Waiting for uninstaller to finish...'
   Wait-Process -Name "unins000*" -Timeout 180 # 3 minutes
 }
 
-Remove-Item $fileManual
+$exists = Test-Path -Path $fileManual -ea 0
+if ($exists){
+  Remove-Item $fileManual
+}
 
 # Delete empty installation folder.
-$FolderNotEmpty = Test-Path -Path $fileManualDir\*
-if (-Not $FolderNotEmpty){
-    Remove-Item $fileManualDir
+$empty = -Not (Test-Path -Path $fileManualDir\* -ea 0)
+if ($empty){
+  Remove-Item $fileManualDir
 }
